黃威誠.新增: Roomtype和Room的interface
