+++ src/main/java/com/booking/dao/booking/RoomDao.java
@@ -2,127 +2,71 @@ package com.booking.dao.booking;
 
 import java.util.List;
 
-import org.hibernate.Session;
-import org.hibernate.query.Query;
-
 import com.booking.bean.booking.Room;
-import com.booking.bean.booking.Roomtype;
 import com.booking.utils.util.DaoResult;
 
-public class RoomDao {
-	private final Session session;
-	
-	public RoomDao(Session session) {
-		this.session = session;
-	}
-	
+public interface RoomDao {
+
 	/**
 	 * 查詢所有房間
 	 * @return
 	 */
-	public DaoResult<List<Room>> getRoomAll() {
-		String hql = "From Room r ORDER BY r.roomNumber";
-		Query<Room> query = session.createQuery(hql, Room.class);
-		query.setFirstResult(0);
-		query.setMaxResults(10);
-		List<Room> rooms = query.getResultList();
-		return DaoResult.create(rooms).setSuccess(rooms != null);
-	}
-	
+	DaoResult<List<Room>> getRoomAll();
+
 	/**
 	 * 根據roomId查找單筆房間
 	 * @param roomId
 	 * @return
 	 */
-	public DaoResult<Room> getRoomById(Integer roomId) {
-		Room room = session.get(Room.class, roomId);		
-		return DaoResult.create(room).setSuccess(room != null); 
-	}
-	
+	DaoResult<Room> getRoomById(Integer roomId);
+
 	/**
 	 * 根據roomtypeId查找多筆房間
 	 * @param roomtypeId
 	 * @return
 	 */
-	public DaoResult<List<Room>> getRoomsByRoomtypeId(Integer roomtypeId) {
-		Roomtype roomtype = session.get(Roomtype.class, roomtypeId);
-		List<Room> rooms = roomtype.getRooms();
-		return DaoResult.create(rooms).setSuccess(rooms != null);
-	}
+	DaoResult<List<Room>> getRoomsByRoomtypeId(Integer roomtypeId);
 
 	/**
 	 * 添加空房間
 	 * @param room
 	 * @return
 	 */
-	public DaoResult<?> addRoom(Room room) {
-		session.persist(room);
-		Integer roomId = room.getRoomId();
-		return DaoResult.create().setGeneratedId(roomId).setSuccess(roomId != null); 
-	}
-	
+	DaoResult<?> addRoom(Room room);
+
 	/**
 	 * 根據roomId刪除房間
 	 * @param roomId
 	 * @return
 	 */
-	public DaoResult<?> removeRoomById(Integer roomId) {
-		Room room = session.get(Room.class, roomId);
-		if(room != null) {
-			session.remove(room);
-			return DaoResult.create().setSuccess(true);
-		}
-		return DaoResult.create().setSuccess(false);
-	}
-	
+	DaoResult<?> removeRoomById(Integer roomId);
+
 	/**
 	 * 根據roomtypeId刪除多筆房間
 	 * @param roomtypeId
 	 * @return
 	 */
-	public DaoResult<?> removeRoomsByRoomtypeId(Integer roomtypeId) {
-		Roomtype roomtype = session.get(Roomtype.class, roomtypeId);
-		if(roomtype != null) {
-			session.remove(roomtype);
-			return DaoResult.create().setSuccess(true);
-		}
-		return DaoResult.create().setSuccess(false);
-	}
+	DaoResult<?> removeRoomsByRoomtypeId(Integer roomtypeId);
 
 	/**
 	 * 更新房間
 	 * @param room
 	 * @return
 	 */
-	public DaoResult<?> updateRoom(Room room) {
-		Room mergeRoom = session.merge(room);
-		return DaoResult.create().setSuccess(mergeRoom != null);
-	}
-	
+	DaoResult<?> updateRoom(Room room);
+
 	/**
 	 * 刪除房間減一房型數量
 	 * @param roomtypeId
 	 * @return
 	 */
-	public DaoResult<?> decrementRoomtypeQuantity(Integer roomId) {
-		Room room = session.get(Room.class, roomId);
-		Roomtype roomtype = room.getRoomtype();
-		Integer quantity = roomtype.getRoomtypeQuantity();
-		roomtype.setRoomtypeQuantity(--quantity);
-		
-		return DaoResult.create().setSuccess(true);
-	}
-	
+	DaoResult<?> decrementRoomtypeQuantity(Integer roomId);
+
 	/**
 	 * 刪除房間加一房型數量
 	 * @param roomtypeId
 	 * @return
 	 */
-	public DaoResult<?> inecrementRoomtypeQuantity(Integer roomtypeId) {
-		Roomtype roomtype = session.get(Roomtype.class, roomtypeId);
-		Integer roomtypeQuantity = roomtype.getRoomtypeQuantity();
-		roomtype.setRoomtypeQuantity(++roomtypeQuantity);
-		
-		return DaoResult.create().setSuccess(true);
-	}
+	DaoResult<?> inecrementRoomtypeQuantity(Integer roomtypeId);
+
 }
